// src/types/config.ts
export interface CommentConfig {
  enable: boolean; // 是否启用评论功能
  system?: "twikoo" | "giscus"; // 评论系统选择
  twikoo?: TwikooConfig;
  giscus?: GiscusConfig;
}

export interface GiscusConfig {
  repo: string;
  repoId: string;
  category: string;
  categoryId: string;
  mapping: string;
  strict: string;
  reactionsEnabled: string;
  emitMetadata: string;
  inputPosition: string;
  theme: string;
  lang: string;
  loading: string;
}

interface TwikooConfig {
  envId: string;
  region?: string;
  lang?: string;
}